\name{AggregateData}
\alias{AggregateData}
\alias{AggregateData.default}
\title{Aggregate Data}
\description{
 Aggregate data set
}
\usage{
AggregateData(data, var, grp, stat = c("mean"), append = FALSE)
}
\arguments{
  \item{data}{data frame}
  \item{var}{character string}
  \item{grp}{character string}
  \item{stat}{statistics to be computed}
  \item{append}{logical}
}
\author{Alaine A. Gulles}
\keyword{dataset}
