\name{allFactors}
\alias{allFactors}
\alias{allFactors.default}
\title{allFactors}
\description{Print all factors of an integer}
\usage{
allFactors(x)
}
\arguments{
  \item{x}{an integer}
}
\references{
   http://stackoverflow.com/questions/6424856/r-function-for-returning-all-factors
}
\author{
  Chase, Ramnath
}
\keyword{misc}
