\name{allPossibleReg}
\alias{allPossibleReg}
\alias{allPossibleReg.default}
\title{All Possible Regression}
\description{All Possible Regression}
\usage{
allPossibleReg(data, depVar, indepVar, selectionStat = "mallow", constant = TRUE)
}
\arguments{
  \item{data}{a}
  \item{depVar}{a}
  \item{indepVar}{a}
  \item{selectionStat}{a}
  \item{constant}{a}
}
\author{
 RIZAMorantte
}
\keyword{analysis}
