\name{AOVTest}
\alias{AOVTest}
\alias{AOVTest.default}
\title{Analysis of Variance}
\description{
 Perform analysis of variance with factors treated as fixed factor.
}
\usage{
AOVTest(data, design, respvar, factor1, factor2 = NULL, factor3 = NULL, factor4 = NULL, rep1 = NULL, rep2 = NULL, set = NULL, descriptive = FALSE, normality = FALSE, homogeneity = FALSE, pwTest = NULL, pwVar = NULL, contrastOption = NULL, sig = 0.05, outputPath = NULL)
}
\arguments{
  \item{data}{a}
  \item{design}{a}
  \item{respvar}{a}
  \item{factor1}{a}
  \item{factor2}{a}
  \item{factor3}{a}
  \item{factor4}{a}
  \item{rep1}{a}
  \item{rep2}{a}
  \item{set}{a}
  \item{descriptive}{a}
  \item{normality}{a}
  \item{homogeneity}{a}
  \item{pwTest}{a}
  \item{pwVar}{a}
  \item{contrastOption}{a}
  \item{sig}{a}
  \item{outputPath}{a}
}
\author{author1}
\keyword{ analysis }
