\name{AppendData}
\alias{AppendData}
\alias{AppendData.default}
\title{Append Data}
\description{
 Append two data sets
}
\usage{
AppendData(MasterData, TransactionData, byMaster, byTransact, MasterVarKeep = NULL, TransactVarKeep = NULL)
}
\arguments{
  \item{MasterData}{a data frame}
  \item{TransactionData}{data frame}
  \item{byMaster}{character string or vector}
  \item{byTransact}{character string or vector}
  \item{MasterVarKeep}{character string or vector}
  \item{TransactVarKeep}{character string or vector}
}
\author{Alaine A. Gulles}
\keyword{dataset}
