\name{buildAOVModel}
\alias{buildAOVModel}
\alias{buildAOVModel.default}
\title{Build AOV Model}
\description{Model building}
\usage{
buildAOVModel(design, respvar, factor1, factor2 = NULL, factor3 = NULL, factor4 = NULL, rep1 = NULL, rep2 = NULL, set = NULL)
}
\arguments{
  \item{design}{a}
  \item{respvar}{a}
  \item{factor1}{a}
  \item{factor2}{a}
  \item{factor3}{a}
  \item{factor4}{a}
  \item{rep1}{a}
  \item{rep2}{a}
  \item{set}{a}
}
\author{
  Alaine A. Gulles
}
\keyword{misc}
