\name{CombineFactorLevels}
\alias{CombineFactorLevels}
\alias{CombineFactorLevels.default}
\title{Combine Factor Levels}
\description{description1}
\usage{
CombineFactorLevels(data, concatVar, targetName = NULL)
}
\arguments{
  \item{data}{a data frame}
  \item{concatVar}{string}
  \item{targetName}{string}
}
\author{
 AAG
}
\keyword{keyword}
