\name{CompareMedian}
\alias{CompareMedian}
\alias{CompareMedian.default}
\title{title1}
\description{description1}
\usage{
CompareMedian(data, varX, varY = NULL, grp = NULL, procedure = c("one", "paired", "independent"), altHypo = c("two.sided", "less", "greater"), hypoValue = 0, confInt = FALSE, confLevel = 0.95)
}
\arguments{
  \item{data}{data frame}
  \item{varX}{string}
  \item{varY}{string}
  \item{grp}{string}
  \item{procedure}{string}
  \item{altHypo}{string}
  \item{hypoValue}{numeric}
  \item{confInt}{logical}
  \item{confLevel}{numeric}
}
\author{author1}
\keyword{ median }
