\name{CompareProportion}
\alias{CompareProportion}
\alias{CompareProportion.default}
\title{Compare Proportion}
\description{Compare Proportion}
\usage{
CompareProportion(data, varX, varY = NULL, grp = NULL, testVal = NULL, procedure = c("one", "paired", "independent"), alternative = c("two.sided", "greater", "less"), CI = FALSE, confLevel = 0.95)
}
\arguments{
  \item{data}{a}
  \item{varX}{a}
  \item{varY}{a}
  \item{grp}{a}
  \item{testVal}{a}
  \item{procedure}{a}
  \item{alternative}{a}
  \item{CI}{a}
  \item{confLevel}{a}
}
\author{aagulles}
\keyword{ analysis }
