\name{confint2}
\alias{confint2}
\title{confint2}
\description{
  confint2
}
\usage{
confint2(object, parm, level = 0.95, method = c("asymptotic", "profile"), ...)
}
\arguments{
  \item{object}{a}
  \item{parm}{a}
  \item{level}{a}
  \item{method}{a}
  \item{\dots}{a}
}
\author{
 from package nlrwr
}
\keyword{analysis}
