\name{ConstructAOVTable}
\alias{ConstructAOVTable}
\alias{ConstructAOVTable.default}
\alias{ConstructAOVTable.summary.aov}
\alias{ConstructAOVTable.summary.aovlist}
\title{Construct the ANOVA Table}
\description{description1}
\usage{
ConstructAOVTable(aovTable, computeTotal = TRUE)
}
\arguments{
  \item{aovTable}{R object}
  \item{computeTotal}{logical}
}
\author{AAGulles of International Rice Research Institute}
\keyword{ table }
