\name{ContrastCompute}
\alias{ContrastCompute}
\alias{ContrastCompute.default}
\title{Constrast Compute}
\description{Constrast Compute}
\usage{
ContrastCompute(data, aovTable, mymodel, mymodel2, contrast.option)
}
\arguments{
  \item{data}{data frame}
  \item{aovTable}{R object}
  \item{mymodel}{model}
  \item{mymodel2}{model}
  \item{contrast.option}{option}
}
\author{author1}
\keyword{analysis}
