\name{Crosstab}
\alias{Crosstab}
\alias{Crosstab.default}
\title{title1}
\description{
  description1
}
\usage{
  Crosstab(data, rowVar, colVar, freqVar = NULL, obsFreq = TRUE, expFreq = FALSE, stdresid = FALSE, totPercent = FALSE, rowPercent = FALSE, colPercent = FALSE, chisq = FALSE, phi = FALSE, cramersv = FALSE, contingency = FALSE)
}
\arguments{
  \item{data}{data frame}
  \item{rowVar}{a character string}
  \item{colVar}{a character string}
  \item{freqVar}{a character string}
  \item{obsFreq}{logical}
  \item{expFreq}{logical}
  \item{stdresid}{logical}
  \item{totPercent}{logical}
  \item{rowPercent}{logical}
  \item{colPercent}{logical}
  \item{chisq}{logical}
  \item{phi}{logical}
  \item{cramersv}{logical}
  \item{contingency}{logical}
}
\author{
  AAG
}
\keyword{keyword1}
