\name{css}
\alias{css}
\alias{css.default}
\alias{css.data.frame}
\title{Corrected Sum of Squares}
\description{
  Computes the corrected sum of squares.
}
\usage{
  css(x, na.rm = TRUE)
}
\arguments{
  \item{x}{a numeric vector}
  \item{na.rm}{logical value indicating whether \code{NA}
               values should be stripped before the 
	       computation proceeds.}
}
\author{Alaine A. Gulles}
\keyword{univar}

