\name{cv}
\alias{cv}
\alias{cv.default}
\alias{cv.data.frame}
\alias{coefVar}
\title{Coefficient of Variation}
\description{
  Generate the coefficient of variation based on a sample data.
}
\usage{
  cv(x, na.rm = TRUE)
  coefVar(x, na.rm = TRUE)
}
\arguments{
  \item{x}{An \R object}
  \item{na.rm}{a logical value indicating whether \code{NA} should
               be stripped before the computation proceeds.}
}
\author{Alaine A. Gulles}
\keyword{univar}
