\name{data.Normalization}
\alias{data.Normalization}
\title{Data Normalization}
\description{
Types of variable normalization formulas
}
\usage{
data.Normalization(x, type = "n0")
}
\arguments{
  	\item{x}{
	vector, matrix or dataset to be tested
	}
  	\item{type}{
	type of normalization
	}
}
\references{clusterSim package.pdf}
\author{Marek Walesiak and Andrzej Dudek}
\keyword{ normal }
