\name{DataAttribute}
\alias{DataAttribute}
\alias{DataAttribute.default}
\title{Dataset Attribute}
\description{
  Determines the variables in the data frame and the variable type.
}
\usage{
  DataAttribute(data)
}
\arguments{
  \item{data}{a data frame or a character string indicating the name
of a data frame that will be used}
}
\value{
  A data frame.
}
\author{Alaine A. Gulles}
\keyword{attributes}
