\name{DataTransformation}
\alias{DataTransformation}
\alias{DataTransformation.default}
\title{title1}
\description{description1}
\usage{
DataTransformation(data, var, transformation = c("log", "ln", "sqrt", "exp", "power", "standardize"), targetName = "newVar")
}
\arguments{
  \item{data}{data frame}
  \item{var}{string}
  \item{transformation}{method}
  \item{targetName}{string}
}
\author{
  AAG
}
\keyword{keyword1}
