\name{designAlphaLattice}
\alias{designAlphaLattice}
\alias{designAlphaLattice.default}
\title{Alpha Lattice}
\description{
description1
}
\usage{
designAlphaLattice(generate, blksize, r = 2, trial = 1, rowPerBlk, rowPerRep, numFieldRow, serpentine = FALSE, file = NULL) 
}
\arguments{
  \item{generate}{generate}
  \item{blksize}{block size}
  \item{r}{replicate}
  \item{trial}{trial}
  \item{rowPerBlk}{numeric}
  \item{rowPerRep}{numeric}
  \item{numFieldRow}{numeric}
  \item{serpentine}{logical}
  \item{file}{filename}
}
\author{author1}
\keyword{ design }
