\name{designAugmentedAlpha}
\alias{designAugmentedAlpha}
\alias{designAugmentedAlpha.default}
\title{Randomization for Augmented Alpha Lattice Design}
\description{Randomization for Augmented Alpha Lattice Design}
\usage{
designAugmentedAlpha(numCheck, numNew, trmtName = NULL, blksize, r = 2, trial = 1, rowPerBlk, rowPerRep, numFieldRow, serpentine = FALSE, trmtLabel = NULL, checkTrmt = NULL, newTrmt = NULL, file = NULL)
}
\arguments{
  \item{numCheck}{numeric}
  \item{numNew}{a}
  \item{trmtName}{a}
  \item{blksize}{a}
  \item{r}{a}
  \item{trial}{a}
  \item{rowPerBlk}{a}
  \item{rowPerRep}{a}
  \item{numFieldRow}{a}
  \item{serpentine}{a}
  \item{trmtLabel}{a}
  \item{checkTrmt}{a}
  \item{newTrmt}{a}
  \item{file}{a}
}
\author{Violeta I Bartolome, Alaine A Gulles}
\keyword{randomization}
