\name{designAugmentedLSD}
\alias{designAugmentedLSD}
\alias{designAugmentedLSD.default}
\title{Augmented Design for Latin Square}
\description{
 Randomization for augmented design in Latin Square
}
\usage{
designAugmentedLSD(numCheck, numNew, trmtName = NULL, trial = 1, numFieldRow, serpentine = FALSE, trmtLabel = NULL, checkTrmt = NULL, newTrmt = NULL, file = NULL)
}
\arguments{
  \item{numCheck}{a}
  \item{numNew}{a}
  \item{trmtName}{a}
  \item{trial}{a}
  \item{numFieldRow}{a}
  \item{serpentine}{a}
  \item{trmtLabel}{a}
  \item{checkTrmt}{a}
  \item{newTrmt}{a}
  \item{file}{a}
}
\author{
 Alaine A. Gulles
}
\keyword{ randomization }
