\name{designAugmentedRCB}
\alias{designAugmentedRCB}
\alias{designAugmentedRCB.default}
\title{Randomization for Augmented Randomized Complete Block Design}
\description{Randomization for Augmented Randomized Complete Block Design}
\usage{
designAugmentedRCB(numCheck, numNew, trmtName = NULL, r = 2, trial = 1, rowPerBlk, numFieldRow, serpentine = FALSE, trmtLabel = NULL, checkTrmt = NULL, newTrmt = NULL, file = NULL)
}
\arguments{
  \item{numCheck}{numeric}
  \item{numNew}{numeric}
  \item{trmtName}{NULL or character string}
  \item{r}{numeric}
  \item{trial}{numeric}
  \item{rowPerBlk}{numeric}
  \item{numFieldRow}{numeric}
  \item{serpentine}{boolean}
  \item{trmtLabel}{string}
  \item{checkTrmt}{string}
  \item{newTrmt}{string}
  \item{file}{a}
}
\author{Violeta I. Bartolome, Alaine Gulles}
\keyword{randomization}
\keyword{augmented}
\keyword{rcbd}
