\name{designAugmentedRowColumn}
\alias{designAugmentedRowColumn}
\alias{designAugmentedRowColumn.default}
\title{Randomization for Augmented Row-Column Design}
\description{Randomization for Augmented Row-Column Design}
\usage{
designAugmentedRowColumn(numCheck, numNew, trmtName = NULL, r = 2, trial = 1, rowblkPerRep, rowPerRowblk, numFieldRow, serpentine = FALSE, trmtLabel = NULL, checkTrmt = NULL, newTrmt = NULL, file = NULL)
}
\arguments{
  \item{numCheck}{a}
  \item{numNew}{a}
  \item{trmtName}{a}
  \item{r}{a}
  \item{trial}{a}
  \item{rowblkPerRep}{a}
  \item{rowPerRowblk}{a}
  \item{numFieldRow}{a}
  \item{serpentine}{a}
  \item{trmtLabel}{a}
  \item{checkTrmt}{a}
  \item{newTrmt}{a}
  \item{file}{a}
}
\author{VIBartolome, AAGulles}
\keyword{Randomization}
