\name{designBIBD}
\alias{designBIBD}
\alias{designBIBD.default}
\title{Randomization of Balanced Incomplete Block Design}
\description{Generates randomization for balanced incomplete block design}
\usage{
designBIBD(generate, blkSize, trial = 1, numFieldRow = 1, rowPerBlk = 1, serpentine = FALSE, display = TRUE, file = NULL)
}
\arguments{
  \item{generate}{a list}
  \item{blkSize}{numeric}
  \item{trial}{numeric}
  \item{numFieldRow}{numeric}
  \item{rowPerBlk}{numeric}
  \item{serpentine}{logical}
  \item{display}{logical}
  \item{file}{NULL}
}
\author{Alaine A. Gulles}
\keyword{design}
