\name{designCRD}
\alias{designCRD}
\alias{designCRD.default}
\title{Randomization of Completely Randomized Design}
\description{description1}
\usage{
designCRD(generate, r = 2, trial = 1, numFieldRow = 1, serpentine = FALSE, display = TRUE, file = NULL)
}
\arguments{
  \item{generate}{a list}
  \item{r}{numeric}
  \item{trial}{numeric}
  \item{numFieldRow}{numeric}
  \item{serpentine}{logical}
  \item{display}{logical}
  \item{file}{NULL}
}
\author{Alaine A. Gulles}
\keyword{ design }
