\name{designLatinizedAlpha}
\alias{designLatinizedAlpha}
\alias{designLatinizedAlpha.default}
\title{Latinized Alpha Design}
\description{
  Randomization for latinized Alpha Lattice design
}
\usage{
designLatinizedAlpha(generate, blksize, r = 2, trial = 1, numFieldRow, serpentine = FALSE, file = NULL)
}
\arguments{
  \item{generate}{list}
  \item{blksize}{numeric}
  \item{r}{replicate}
  \item{trial}{numeric}
  \item{numFieldRow}{numeric}
  \item{serpentine}{logical}
  \item{file}{path}
}
\author{AAGulles}
\keyword{randomization}
