\name{designLatinizedRowCol}
\alias{designLatinizedRowCol}
\alias{designLatinizedRowCol.default}
\title{Latinized Row-Column}
\description{
  Randomization for Latinized Row-Column Design
}
\usage{
designLatinizedRowCol(generate, r = 2, trial = 1, rowPerRep, numFieldRow, serpentine = FALSE, file = NULL)
}
\arguments{
  \item{generate}{list}
  \item{r}{numeric}
  \item{trial}{numeric}
  \item{rowPerRep}{numeric}
  \item{numFieldRow}{numeric}
  \item{serpentine}{logical}
  \item{file}{path}
}
\author{AAGulles}
\keyword{Randomization}
