\name{designLSD}
\alias{designLSD}
\alias{designLSD.default}
\title{Randomization for Latin Square Design}
\description{Randomization for Latin Square Design}
\usage{
designLSD(generate, trial = 1, serpentine = FALSE, display = TRUE, file = NULL)
}
\arguments{
  \item{generate}{a list}
  \item{trial}{numeric}
  \item{serpentine}{logical}
  \item{display}{logical}
  \item{file}{NULL}
}
\author{Alaine A. Gulles}
\keyword{ design }
