\name{designPRep}
\alias{designPRep}
\alias{designPRep.default}
\title{Randomization for p-Rep Design}
\description{Randomization for p-Rep Design}
\usage{
designPRep(trmtPerGrp, trmtRepPerGrp, trmtName = NULL, trial = 1, numFieldRow, serpentine = FALSE, trmtLabel = NULL, trmtListPerGrp = NULL, file = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{trmtPerGrp}{a}
  \item{trmtRepPerGrp}{a}
  \item{trmtName}{a}
  \item{trial}{a}
  \item{numFieldRow}{a}
  \item{serpentine}{a}
  \item{trmtLabel}{a}
  \item{trmtListPerGrp}{a}
  \item{file}{a}
}
\author{VIBartolome, AAGulles}
\keyword{Randomization}
