\name{designRCBD}
\alias{designRCBD}
\alias{designRCBD.default}
\title{Randomization for Randomized Complete Block Design}
\description{description1}
\usage{
designRCBD(generate, r = 2, trial = 1, numFieldRow = 1, rowPerBlk = 1, serpentine = FALSE, display = TRUE, file = NULL)
}
\arguments{
  \item{generate}{a list}
  \item{r}{a}
  \item{trial}{a}
  \item{numFieldRow}{a}
  \item{rowPerBlk}{a}
  \item{serpentine}{a}
  \item{display}{a}
  \item{file}{a}
}
\author{Alaine A. Gulles}
\keyword{ design }
