\name{designSplit}
\alias{designSplit}
\alias{designSplit.default}
\title{Randomization of Family of Split Plot Design}
\description{description}
\usage{
designSplit(main, sub, ssub = NULL, sssub = NULL, r = NULL, trial = 1, design = c('crd', 'rcbd', 'lsd'), 
            numFieldRow = 1, rowPerBlk = 1, rowPerMain = 1, rowPerSub = 1, rowPerSubSub = 1, 
            serpentine = FALSE, display = TRUE, file = NULL)
}
\arguments{
  \item{main}{a}
  \item{sub}{a}
  \item{ssub}{a}
  \item{sssub}{a}
  \item{r}{a}
  \item{trial}{a}
  \item{design}{a}
  \item{numFieldRow}{numeric}
  \item{rowPerBlk}{numeric}
  \item{rowPerMain}{numeric}
  \item{rowPerSub}{numeric}
  \item{rowPerSubSub}{numeric} 
  \item{serpentine}{logical}
  \item{display}{a}
  \item{file}{a}
}
\author{Alaine A. Gulles}
\keyword{ design }
