\name{designStrip}
\alias{designStrip}
\alias{designStrip.default}
\title{Randomization for Strip Plot design}
\description{Randomization for Strip Plot design}
\usage{
designStrip(vertical, horizontal, sub = NULL, ssub = NULL, r = 2, trial = 1, numFieldRow, 
                        rowPerMain, rowPerSub, serpentine, display = TRUE, file = NULL)
}
\arguments{
  \item{vertical}{a}
  \item{horizontal}{a}
  \item{sub}{a}
  \item{ssub}{a}
  \item{r}{a}
  \item{trial}{a}
  \item{numFieldRow}{numeric}
  \item{rowPerMain}{numeric}
  \item{rowPerSub}{numeric}
  \item{serpentine}{logical} 
  \item{display}{a}
  \item{file}{a}
}
\author{Alaine A. Gulles}
\keyword{design}
