\name{Discriminant}
\alias{Discriminant}
\alias{Discriminant.default}
\title{Discriminant}
\description{Discriminant Analysis}
\usage{
Discriminant(data, outputPath, vars, grpVar, descriptive = FALSE, doNormalTest = FALSE, doBoxM = FALSE, priorProb = NULL, classifyNew = FALSE, data2)
}
\arguments{
  \item{data}{a}
  \item{outputPath}{a}
  \item{vars}{a}
  \item{grpVar}{a}
  \item{descriptive}{a}
  \item{doNormalTest}{a}
  \item{doBoxM}{a}
  \item{priorProb}{a}
  \item{classifyNew}{a}
  \item{data2}{a}
}
\author{
 RIZAMorantte
}
\keyword{multivariate}
