\name{estMissData}
\alias{estMissData}
\alias{estMissData.default}
\title{Estimate Missing Data}
\description{Estimate Missing Data}
\usage{
estMissData(design, data, respvar, factor1, factor2, factor3, factor4, rep1, rep2)
}
\arguments{
  \item{design}{design}
  \item{data}{a}
  \item{respvar}{a}
  \item{factor1}{a}
  \item{factor2}{a}
  \item{factor3}{a}
  \item{factor4}{a}
  \item{rep1}{a}
  \item{rep2}{a}
}
\author{author1}
\keyword{ estimate }
