\name{FitDistrib}
\alias{FitDistrib}
\alias{FitDistrib.default}
\title{Fit Distribution}
\description{Fit Distribution}
\usage{
FitDistrib(data, outputPath, yVar, byVar = NULL, distrib = c("beta", "exp", "gamma", "geom", "lnorm", "logis", "nbinom", "norm", "pois"), fitMethod = c("mle", "mme", "mge"))
}
\arguments{
  \item{data}{a}
  \item{outputPath}{a}
  \item{yVar}{a}
  \item{byVar}{a}
  \item{distrib}{a}
  \item{fitMethod}{a}
}
\author{
 Rose Imee Zhella A. Morantte
}
\keyword{distribution}
