\name{GenerateBalanceData}
\alias{GenerateBalanceData}
\alias{GenerateBalanceData.default}
\title{Generate Balance Data}
\description{Generate Balance Data}
\usage{
GenerateBalanceData(data, respvar, factor, blk, design)
}
\arguments{
  \item{data}{a}
  \item{respvar}{a}
  \item{factor}{a}
  \item{blk}{a}
  \item{design}{a}
}
\author{author1}
\keyword{ data }
