\name{GoodnessOfFitTest}
\alias{GoodnessOfFitTest}
\alias{GoodnessOfFitTest.default}
\title{Chi-Square Goodness of Fit Test}
\description{Chi-Square Goodness of Fit Test}
\usage{
GoodnessOfFitTest(data, var, colFreq = NULL, expected = NULL)
}
\arguments{
  \item{data}{data frame}
  \item{var}{string}
  \item{colFreq}{string}
  \item{expected}{string}
}
\author{aagulles}
\keyword{ analysis }
