\name{GraphBoxplot}
\alias{GraphBoxplot}
\alias{GraphBoxplot.default}
\title{Graph Boxplot}
\description{GraphBoxplot}
\usage{
GraphBoxplot(data, outputPath, nVar, cVars = NULL, mTitle = NULL, nAxisLab = NULL, cAxisLab = NULL,
                         yMinValue = NULL, yMaxValue = NULL, axisLabelStyle = 1, plotHoriz = FALSE, byVar = NULL, 
                         boxSize = 0.8, boxColor = NULL, boxFillColor = NULL,
                         medLineType = 1, medLineWidth = 3, medColor = NULL, whiskLineType = 2, whiskLineWidth = 1,
                         whiskColor = NULL, outChar = 1, outCharSize = 1, outColor = NULL,
                         boxed = TRUE, multGraphs = FALSE, numRowsGraphs = 1, numColsGraphs = 1, 
                         orientGraphs = c("left-right", "top-bottom"))
}
\arguments{
  \item{data}{a}
  \item{outputPath}{a}
  \item{nVar}{a}
  \item{cVars}{a}
  \item{mTitle}{a}
  \item{nAxisLab}{a}
  \item{cAxisLab}{a}
  \item{yMinValue}{a}
  \item{yMaxValue}{a}
  \item{axisLabelStyle}{a}
  \item{plotHoriz}{a}
  \item{byVar}{a}
  \item{boxSize}{a}
  \item{boxColor}{a}
  \item{boxFillColor}{a}
  \item{medLineType}{a}
  \item{medLineWidth}{a}
  \item{medColor}{a}
  \item{whiskLineType}{a}
  \item{whiskLineWidth}{a}
  \item{whiskColor}{a}
  \item{outChar}{a}
  \item{outCharSize}{a}
  \item{outColor}{a}
  \item{boxed}{a}
  \item{multGraphs}{a}
  \item{numRowsGraphs}{a}
  \item{numColsGraphs}{a}
  \item{orientGraphs}{a}
}
\author{
Rose Imee Zhella A. Morantte
}
\keyword{boxplot}
