\name{GraphHist}
\alias{GraphHist}
\alias{GraphHist.default}
\title{Histogram}
\description{GraphHist}
\usage{
GraphHist(data, outputPath, nVar, mTitle = NULL, yAxisLabs, xAxisLabs, 
              yMinValues, yMaxValues, xMinValues, xMaxValues, 
              useFreq = TRUE, numBins = "Sturges", axisLabelStyle = 1, byVar = NULL, 
              barColor, dispCurve = FALSE, lineType, lineColor, lineWidth,
              boxed = TRUE, multGraphs = FALSE, numRowsGraphs = 1, numColsGraphs = 1, 
              orientGraphs = c("left-right", "top-bottom"))
}
\arguments{
  \item{data}{a}
  \item{outputPath}{a}
  \item{nVar}{a}
  \item{mTitle}{a}
  \item{yAxisLabs}{a}
  \item{xAxisLabs}{a}
  \item{yMinValues}{a}
  \item{yMaxValues}{a}
  \item{xMinValues}{a}
  \item{xMaxValues}{a}
  \item{useFreq}{a}
  \item{numBins}{a}
  \item{axisLabelStyle}{a}
  \item{byVar}{a}
  \item{barColor}{a}
  \item{dispCurve}{a}
  \item{lineType}{a}
  \item{lineColor}{a}
  \item{lineWidth}{a}
  \item{boxed}{a}
  \item{multGraphs}{a}
  \item{numRowsGraphs}{a}
  \item{numColsGraphs}{a}
  \item{orientGraphs}{a}
}
\author{
Rose Imee Zhella A. Morantte
}
\keyword{histogram}
