\name{GraphLine}
\alias{GraphLine}
\alias{GraphLine.default}
\title{GraphLine}
\description{
constructs line graph
}
\usage{
GraphLine(data, outputPath, yVars, xVar = NULL, lineVars = NULL, mTitle = NULL, yAxisLab = NULL, 
                      xAxisLab = NULL, yMinValue = NULL, yMaxValue = NULL, axisLabelStyle = 1,
                      byVar = NULL, errBars = FALSE, typeErrBar = c(NULL, "stdErr", "confInt"), errMult = 1, confLev = NULL, 
                      plotCol = NULL, showLineLabels = FALSE, showLeg = FALSE, 
                      legPos = c("bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right",	"center"), 
                      legTitle = NULL, legCol= 1, boxed = TRUE, linePtTypes, lineTypes,  lineWidths, pointChars, pointCharSizes, 
                      multGraphs = FALSE, numRowsGraphs = 1, numColsGraphs = 1, orientGraphs = c("left-right", "top-bottom"))
}
\arguments{
  \item{data}{a}
  \item{outputPath}{a}
  \item{yVars}{a}
  \item{xVar}{a}
  \item{lineVars}{a}
  \item{mTitle}{a}
  \item{yAxisLab}{a}
  \item{xAxisLab}{a}
  \item{yMinValue}{a}
  \item{yMaxValue}{a}
  \item{axisLabelStyle}{a}
  \item{byVar}{a}
  \item{errBars}{a}
  \item{typeErrBar}{a}
  \item{errMult}{a}
  \item{confLev}{a}
  \item{plotCol}{a}
  \item{showLineLabels}{a}
  \item{showLeg}{a}
  \item{legPos}{a}
  \item{legTitle}{a}
  \item{legCol}{a}
  \item{boxed}{a}
  \item{linePtTypes}{a}
  \item{lineTypes}{a}
  \item{lineWidths}{a}
  \item{pointChars}{a}
  \item{pointCharSizes}{a}
  \item{multGraphs}{a}
  \item{numRowsGraphs}{a}
  \item{numColsGraphs}{a}
  \item{orientGraphs}{a}
}
\author{
Rose Imee Zhella A. Morantte
}
\keyword{linegraph}
