\name{GraphPie}
\alias{GraphPie}
\alias{GraphPie.default}
\title{GraphPie}
\description{GraphPie}
\usage{
GraphPie(data, outputPath, cVar, mTitle = NULL, wVar = NULL,
	byVar = NULL, sliceColors, dispLabels = FALSE, useLabels = c("counts", "pcts", "cats", "pctcats", "countcats"), 
        showLeg = FALSE,legPos = c("bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright",
        "right", "center"), legTitle = NULL, legCol = 1, pieDensity, pieLineAngle, boxed = TRUE,
	multGraphs = FALSE, numRowsGraphs = 1, numColsGraphs = 1,
	orientGraphs = c("left-right", "top-bottom"))
}
\arguments{
  \item{data}{a}
  \item{outputPath}{a}
  \item{cVar}{a}
  \item{mTitle}{a}
  \item{wVar}{a}
  \item{byVar}{a}
  \item{sliceColors}{a}
  \item{dispLabels}{a}
  \item{useLabels}{a}
  \item{showLeg}{a}
  \item{legPos}{a}
  \item{legTitle}{a}
  \item{legCol}{a}
  \item{pieDensity}{a}
  \item{pieLineAngle}{a}
  \item{boxed}{a}
  \item{multGraphs}{a}
  \item{numRowsGraphs}{a}
  \item{numColsGraphs}{a}
  \item{orientGraphs}{a}
}
\author{
 Rose Imee Zhella A. Morantte
}
\keyword{graphpie}
