\name{GraphScatter}
\alias{GraphScatter}
\alias{GraphScatter.default}
\title{GraphScatter}
\description{GraphScatter}
\usage{
GraphScatter(data, outputPath, xVar, yVar, mTitle = NULL, xAxisLab = NULL, 
	yAxisLab = NULL, xMinValue = NULL, xMaxValue = NULL , yMinValue = NULL, yMaxValue = NULL,
	axisLabelStyle = 1, byVar = NULL, boxed = TRUE, pointCol, pointChar, pointCharSize, 
        dispLine = FALSE, lineType = NULL, lineCol = NULL, lineWidth, dispR2P = FALSE, 
        r2PPos = c("bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right",  "center"),
        multGraphs = FALSE, numRowsGraphs = 1, numColsGraphs = 1, orientGraphs = c("left-right", "top-bottom"))
}
\arguments{
  \item{data}{a}
  \item{outputPath}{a}
  \item{xVar}{a}
  \item{yVar}{a}
  \item{mTitle}{a}
  \item{xAxisLab}{a}
  \item{yAxisLab}{a}
  \item{xMinValue}{a}
  \item{xMaxValue}{a}
  \item{yMinValue}{a}
  \item{yMaxValue}{a}
  \item{axisLabelStyle}{a}
  \item{byVar}{a}
  \item{boxed}{a}
  \item{pointCol}{a}
  \item{pointChar}{a}
  \item{pointCharSize}{a}
  \item{dispLine}{a}
  \item{lineType}{a}
  \item{lineCol}{a}
  \item{lineWidth}{a}
  \item{dispR2P}{a}
  \item{r2PPos}{a}
  \item{multGraphs}{a}
  \item{numRowsGraphs}{a}
  \item{numColsGraphs}{a}
  \item{orientGraphs}{a}
}
\author{
Rose Imee Zhella A. Morantte
}
\keyword{scatterplot}
