\name{HeteroskedasticityTest}
\alias{HeteroskedasticityTest}
\alias{HeteroskedasticityTest.default}
\title{Heteroskedasticity Test}
\description{Perform equality of variances using bartlett's and levene's test}
\usage{
  HeteroskedasticityTest(data, var, grp, method = c("bartlett"))
}
\arguments{
  \item{data}{data frame}
  \item{var}{character string}
  \item{grp}{character string}
  \item{method}{character string}
}
\author{
  AAG
}
\keyword{keyword1}
