\name{is.wholenumber}
\alias{is.wholenumber}
\title{Wholenumber}
\description{
  Determines if the input is a wholenumber or not.
}
\usage{
  is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
  \item{x}{an object to be tested}
  \item{tol}{a numerical tolerance}
}
\references{
  FULLREFMAN.PDF
}
\author{R Development Core Team}
\keyword{wholenumber}
