\name{kurtosis}
\alias{kurtosis}
\alias{kurtosis.default}
\alias{kurtosis.data.frame}
\alias{coefKurtosis}
\title{Coefficient of Kurtosis}
\description{
  Computes the coefficient of kurtosis bases from a sample.
}
\usage{
  kurtosis(x, na.rm = TRUE)
  coefKurtosis(x, na.rm = TRUE)
}
\arguments{
  \item{x}{An \R object}
  \item{na.rm}{a logical value indicating whether \code{NA} should
               be stripped before the computation proceeds.}
}
\author{Alaine A. Gulles}
\keyword{univar}
