\name{LinearRegressionAnalysis}
\alias{LinearRegressionAnalysis}
\alias{LinearRegressionAnalysis.default}
\title{Linear Regression Analysis}
\description{Linear Regression Analysis}
\usage{
LinearRegressionAnalysis(data, depVar, indepVar, constant = TRUE, statistics = TRUE, confInt = FALSE, confLevel = 0.95, covMatrix = FALSE, normality = NULL, heteroskedasticity = NULL, autoCorr = FALSE, VIF = FALSE, COOKS = FALSE, leverage = FALSE)
}
\arguments{
  \item{data}{a}
  \item{depVar}{a}
  \item{indepVar}{a}
  \item{constant}{a}	
  \item{statistics}{a}
  \item{confInt}{a}
  \item{confLevel}{a}
  \item{covMatrix}{a}
  \item{normality}{a}
  \item{heteroskedasticity}{a}
  \item{autoCorr}{a}
  \item{VIF}{a}
  \item{COOKS}{a}
  \item{leverage}{a}
}
\author{author1}
\keyword{ analysis }
