\name{LinearRegressionSelection}
\alias{LinearRegressionSelection}
\alias{LinearRegressionSelection.default}
\title{Linear Regression Selection}
\description{Linear Regression Selection}
\usage{
LinearRegressionSelection(data, depVar, indepVar, constant = TRUE, correlate = FALSE, 
                                     selection = "none", selectionStat = NULL,
                                     confInt = FALSE, confLevel = 0.95, 
                                     covMatrix = FALSE, normality = NULL, 
                                     heteroskedasticity = NULL, autoCorr = FALSE, 
                                     VIF = FALSE, COOKS = FALSE, leverage = FALSE)
}
\arguments{
  \item{data}{a}
  \item{depVar}{a}
  \item{indepVar}{a}
  \item{constant}{a}
  \item{correlate}{a}
  \item{selection}{a}
  \item{selectionStat}{a}
  \item{confInt}{a}
  \item{confLevel}{a}
  \item{covMatrix}{a}
  \item{normality}{a}
  \item{heteroskedasticity}{a}
  \item{autoCorr}{a}
  \item{VIF}{a}
  \item{COOKS}{a}
  \item{leverage}{a}
}
\author{
RIZAMorantte
}
\keyword{analysis}
