\name{MAOVTest}
\alias{MAOVTest}
\alias{MAOVTest.default}
\title{MANOVA Test}
\description{Multivariate ANOVA}
\usage{
MAOVTest(data, outputPath, yVars, factorVar, repVar = NULL, testStat = c("Pillai","Wilks","Hotelling-Lawley","Roy"), 
         mymodel, descriptive = FALSE, correlate = FALSE, doNormalTest = FALSE, doBoxM= FALSE, doSphericity = FALSE, saveSSH = FALSE, saveSSE = FALSE)
}
\arguments{
  \item{data}{a}
  \item{outputPath}{a}
  \item{yVars}{a}
  \item{factorVar}{a}
  \item{repVar}{a}
  \item{testStat}{a}
  \item{mymodel}{a}
  \item{descriptive}{a}
  \item{correlate}{a}
  \item{doNormalTest}{a}
  \item{doBoxM}{a}
  \item{doSphericity}{a}
  \item{saveSSH}{a}
  \item{saveSSE}{a}
}
\author{
 RIZAM
}
\keyword{ analysis }
