\name{MDS}
\alias{MDS}
\alias{MDS.default}
\title{MDS}
\description{MDS}
\usage{
MDS(data, outputPath, inputType = c("raw", "distMat"), vars = NULL, idVar, 
    type = c("Classical", "Nonmetric"), 
    distClass = c("Euclidean", "Maximum", "Manhattan", "Canberra", "Minkowski"),
    distNonmet = c("Manhattan", "Euclidean", "Canberra", "Bray", "Kulczynski", 
                   "Jaccard", "Gower", "AltGower", "Morisita", "Horn", 
                   "Mountford", "Raup", "Binomial", "Chao", "Cao"), 
    dimnum = 2, useIdVar = FALSE, pChars, pSizes, pCol, showLeg = FALSE, legTitle, 
    legPos = "bottomright", legNcol = 1, descriptive = FALSE, correlate = FALSE)
}
\arguments{
  \item{data}{a}
  \item{outputPath}{a}
  \item{inputType}{a}
  \item{vars}{a}
  \item{idVar}{a}
  \item{type}{a}
  \item{distClass}{a}
  \item{distNonmet}{a}
  \item{dimnum}{a}
  \item{useIdVar}{a}
  \item{pChars}{a}
  \item{pSizes}{a}
  \item{pCol}{a}
  \item{showLeg}{a}
  \item{legTitle}{a}
  \item{legPos}{a}
  \item{legNcol}{a}
  \item{descriptive}{a}
  \item{correlate}{a}
}
\author{
RIZAMorantte
}
\keyword{analysis}
