\name{mdsGraph}
\alias{mdsGraph}
\alias{mdsGraph.default}
\title{mdsGraph}
\description{mdsGraph}
\usage{
mdsGraph(data, idVar, outputPath, pCharName = NULL, pColName, grpLevels = NULL, 
         idData, dimnum, fit, type, useIdVar = FALSE, pChars, pSizes, pCol, 
         showLeg, legTitle, legPos, legNcol)
}
\arguments{
  \item{data}{a}
  \item{idVar}{a}
  \item{outputPath}{a}
  \item{pCharName}{a}
  \item{pColName}{a}
  \item{grpLevels}{a}
  \item{idData}{a}
  \item{dimnum}{a}
  \item{fit}{a}
  \item{type}{a}
  \item{useIdVar}{a}
  \item{pChars}{a}
  \item{pSizes}{a}
  \item{pCol}{a}
  \item{showLeg}{a}
  \item{legTitle}{a}
  \item{legPos}{a}
  \item{legNcol}{a}
}
\author{
  RIZAMorantte
}
\keyword{analysis}
