\name{modalValue}
\alias{modalValue}
\alias{modalValue.default}
\alias{modalValue.data.frame}
\title{Modal Value}
\description{ 
  Determines the most frequent observations in the dataset.
}
\usage{
  modalValue(x, na.rm = TRUE)
}
\arguments{
  \item{x}{An \R object}
  \item{na.rm}{a logical value indicating whether \code{NA} should
               be stripped before the computation proceeds.}
}
\author{Alaine A. Gulles}
\keyword{univar}
