\name{NonLinearRegression}
\alias{NonLinearRegression}
\alias{NonLinearRegression.default}
\title{Non-Linear Regeression}
\description{
   Non-Linear Regression
}
\usage{
NonLinearRegression(data, depVar, indepVar, statistics = FALSE, model, modelExp = NULL, startValues = NULL, normality = FALSE, covMatrix = FALSE, confInterval = FALSE, confLevel = 0.95)
}
\arguments{
  \item{data}{a}
  \item{depVar}{a}
  \item{indepVar}{a}
  \item{statistics}{a}
  \item{model}{a}
  \item{modelExp}{a}
  \item{startValues}{a}
  \item{normality}{a}
  \item{covMatrix}{a}
  \item{confInterval}{a}
  \item{confLevel}{a}
}
\author{
  RIZAMorantte
}
\keyword{analysis}
