\name{pairwiseAmong}
\alias{pairwiseAmong}
\alias{pairwiseAmong.default}
\title{Pairwise Among}
\description{Pairwise Among}
\usage{
pairwiseAmong(data, respvar, typeTest, trmt, dfError, MSError, siglevel)
}
\arguments{
  \item{data}{a}
  \item{respvar}{a}
  \item{typeTest}{a}
  \item{trmt}{a}
  \item{dfError}{a}
  \item{MSError}{a}
  \item{siglevel}{a}
}
\author{a}
\keyword{ keyword }
