\name{pairwiseComparison}
\alias{pairwiseComparison}
\alias{pairwiseComparison.default}
\title{Pairwise Comparison}
\description{Pairwise Comparison}
\usage{
pairwiseComparison(aovTable, design, effect, data, respvar, pwTest = NULL, siglevel = 0.05)
}
\arguments{
  \item{aovTable}{a}
  \item{design}{a}
  \item{effect}{a}
  \item{data}{a}
  \item{respvar}{a}
  \item{pwTest}{a}
  \item{siglevel}{a}
}
\author{author1}
\keyword{ analysis }
