\name{pairwiseWithin}
\alias{pairwiseWithin}
\alias{pairwiseWithin.default}
\title{Pairwise Within}
\description{Pairwise Within}
\usage{
pairwiseWithin(data, respvar, typeTest, nobs1, nobs2, dfError, MSError, f1, f2, f3 = NULL, siglevel)
}
\arguments{
  \item{data}{a}
  \item{respvar}{a}
  \item{typeTest}{a}
  \item{nobs1}{a}
  \item{nobs2}{a}
  \item{dfError}{a}
  \item{MSError}{a}
  \item{f1}{a}
  \item{f2}{a}
  \item{f3}{a}
  \item{siglevel}{a}
}
\author{a}
\keyword{util}
