\name{plotDiscDist}
\alias{plotDiscDist}
\alias{plotDiscDist.default}
\title{plotDiscDist}
\description{plotDiscDist}
\usage{
plotDiscDist(data, distr, para, ...)
}
\arguments{
  \item{data}{a}
  \item{distr}{a}
  \item{para}{a}
  \item{\dots}{a}
}
\author{
  Rose Imee Zhella A. Morantte
}
\keyword{distribution}
