\name{PredictedValues}
\alias{PredictedValues}
\alias{PredictedValues.default}
\alias{PredictedValues.aovlist}
\title{title1}
\description{description1}
\usage{
PredictedValues(object, errorTerm = NULL)
}
\arguments{
  \item{object}{a}
  \item{errorTerm}{a}
}
\author{author1}
\keyword{ prediction }
