\name{print}
\alias{print.powerTransform}
\alias{print.summary.powerTransform}
\title{print}
\description{description1}
\usage{
\method{print}{powerTransform}(x, ...)
\method{print.summary}{powerTransform}(x, digits = 4, ...)
}
\arguments{
  \item{x}{object}
  \item{digits}{number}
  \item{\dots}{additional option}
}
\author{car package}
\keyword{print}
