\name{printBIBDPairwise}
\alias{printBIBDPairwise}
\alias{printBIBDPairwise.default}
\title{Print BIBD Pairwise}
\description{
  Print the pairwise comparison of the Balanced Incomplete Block Design
}
\usage{
printBIBDPairwise(object, dfError, MSError, method, trmtLabel = NULL)
}
\arguments{
  \item{object}{an R object}
  \item{dfError}{dfError}
  \item{MSError}{MSError}
  \item{method}{method of pairwise to be printed}
  \item{trmtLabel}{treatment label}
}
\author{
 AAG
}
\keyword{misc}
