\name{printCorrMatrix}
\alias{printCorrMatrix}
\alias{printCorrMatrix.default}
\title{Print Correlation Matrix}
\description{
  Print the correlation matrix.
}
\usage{
  printCorrMatrix(coef, pval = NULL, n)
}
\arguments{
  \item{coef}{a table of the correlation coefficient}
  \item{pval}{a table of the pvalue}
  \item{n}{a table of number of observation}
}
\author{Alaine A. Gulles}
\keyword{print}
