\name{printDataFrame}
\alias{printDataFrame}
\alias{printDataFrame.default}
\title{Print Data Frames}
\description{
  Print a data frame.
}
\usage{
  printDataFrame(dataFrame, border = TRUE, digits = NULL)
}
\arguments{
  \item{dataFrame}{an object of class data.frame}
  \item{border}{a logical value indicating whether a border between
                the variable name and the data will be displayed}
  \item{digits}{NULL}
}
\author{Alaine A. Gulles of International Rice Research Institute}
\keyword{print}
\keyword{data.frame}
