\name{printLayout}
\alias{printLayout}
\alias{printLayout.default}
\title{print layout}
\description{print layout}
\usage{
printLayout(trmt, plotNum, RowLabel = NULL, ColLabel = NULL, title = NULL)
}
\arguments{
  \item{trmt}{a}
  \item{plotNum}{a}
  \item{RowLabel}{a}
  \item{ColLabel}{a}
  \item{title}{a}
}
\author{
  Alaine A. Gulles
}
\keyword{misc}
