\name{SignificantEffect}
\alias{SignificantEffect}
\alias{SignificantEffect.default}
\title{title1}
\description{description1}
\usage{
SignificantEffect(aovTable, alpha = 0.05)
}
\arguments{
  \item{aovTable}{R object}
  \item{alpha}{level of significance}
}
\author{author1}
\keyword{ pairwise }% __ONLY ONE__ keyword per line
