\name{skewness}
\alias{skewness}
\alias{skewness.default}
\alias{skewness.data.frame}
\alias{skew}
\title{Coefficient of Skewness}
\description{
  Generate the coefficient of skewness based on a sample.
}
\usage{
  skewness(x, na.rm = TRUE)
  skew(x, na.rm = TRUE)
}
\arguments{
  \item{x}{An \R object}
  \item{na.rm}{a logical value indicating whether \code{NA} should
               be stripped before the computation proceeds.}
}
\author{Alaine A. Gulles}
\keyword{univar}
