\name{stdkurtosis}
\alias{stdkurtosis}
\alias{stdkurtosis.default}
\alias{stdkurtosis.data.frame}
\alias{se.kurtosis}
\title{Standard Error of Kurtosis}
\description{
  Generate the standard error of the kurtosis.
}
\usage{
  stdkurtosis(x, na.rm = TRUE)
  se.kurtosis(x, na.rm = TRUE)
}
\arguments{
  \item{x}{An \R object}
  \item{na.rm}{a logical value indicating whether \code{NA} should
               be stripped before the computation proceeds.}
}
\author{Alaine A. Gulles}
\keyword{univar}
