\name{stdmean}
\alias{stdmean}
\alias{stdmean.default}
\alias{stdmean.data.frame}
\alias{se.mean}
\title{Standard Error of the Mean}
\description{
  Generate the standard error of the sample mean.
}
\usage{
  stdmean(x, na.rm = TRUE)
  se.mean(x, na.rm = TRUE)
}
\arguments{
  \item{x}{An \R object}
  \item{na.rm}{a logical value indicating whether \code{NA} should
               be stripped before the computation proceeds.}
}
\author{Alaine A. Gulles}
\keyword{univar}

