\name{stdskewness}
\alias{stdskewness}
\alias{stdskewness.default}
\alias{stdskewness.data.frame}
\alias{se.skewness}
\title{Standard Error of Skewness}
\description{
  Generate the standard error of skewness.
}
\usage{
  stdskewness(x, na.rm = TRUE)
  se.skewness(x, na.rm = TRUE)
}
\arguments{
  \item{x}{An \R object}
  \item{na.rm}{a logical value indicating whether \code{NA} should
               be stripped before the computation proceeds.}
}
\author{Alaine A. Gulles}
\keyword{univar}
