\name{ToLong}
\alias{ToLong}
\alias{ToLong.default}
\title{Reshape to Long format}
\description{Reshape from wide to long format}
\usage{
ToLong(data, varying, timevar, vnames, idvar, label = NULL)
}
\arguments{
  \item{data}{a data frame}
  \item{varying}{character string}
  \item{timevar}{character string}
  \item{vnames}{character string}
  \item{idvar}{character string}
  \item{label}{character string}
}
\author{AAGulles}
\keyword{data}
