\name{ucss}
\alias{ucss}
\alias{ucss.default}
\alias{ucss.data.frame}
\title{Uncorrected Sum of Squares}
\description{
  Generate the uncorrected sum of squares.
}
\usage{
  ucss(x, na.rm = TRUE)
}
\arguments{
  \item{x}{An \R object}
  \item{na.rm}{a logical value indicating whether \code{NA} should
               be stripped before the computation proceeds.}
}
\author{Alaine A. Gulles}
\keyword{univar}

